Rework OpenACC Fortran DO loop initialization
Fortran DO loops on arrays with non-constant bounds (like a(lo:hi)) need
special setup code to compute the bounds and offsets for the iteration. In
an OpenACC region containing multiple loops, this used to be done in a block
of code at the start of the region for all of the loops. But the upcoming
kernels conversion expects this kind of setup code to immediately precede
the corresponding loop, and variables are not mapped correctly otherwise.
This patch separates out the initialization part for each loop and places it
immediately before the loop.

    gcc/fortran/
    * trans-openmp.c (gfc_privatize_nodesc_array_clauses): Renamed from
    gfc_privatize_nodesc_arrays, initialization part factored out to...
    (gfc_reinitialize_privatized_arrays): ... this new function, called...
    (gfc_trans_omp_do): ... from here for OpenACC loops.

    libgomp/
    * testsuite/libgomp.oacc-fortran/initialize_kernels_loops.f90: New test.